requirement.txt
django==4.1.1
pip==21.2.4
django-js-asset==2.0.0
mysqlclient==2.2.0
django-tinymce==3.5.0
django-treebeard==4.7
django-jazzmin==2.6.0
pillow==9.2.0
django-countries==7.4.2
six==1.16.0
django-six==1.0.4